move executor to runtime
