+++ datafusion/execution/src/lib.rs
@@ -22,11 +22,14 @@
 
 pub mod cache;
 pub mod config;
+pub mod cross_rt_stream;
+pub mod dedicated_executor;
 pub mod disk_manager;
+pub mod io_object_store;
 pub mod memory_pool;
 pub mod object_store;
 pub mod runtime_env;
-mod stream;
+pub mod stream;
 mod task;
 
 pub mod registry {
